Abstract over prime fields and Galois field extensions
- Factorization can now be performed over Galois fields
- Automatically upgrade to larger Galois fields when field
  is too small for sampling during factorization
- Add Z2 prime field with elements 0 and 1
- Make sure finite field to_integer does not truncate

+++ src/poly/polynomial.rs
@@ -1790,7 +1790,7 @@ impl<F: Ring, E: Exponent> MultivariatePolynomial<F, E, LexOrder> {
 
             c2.exponents = c
                 .exponents_iter()
-                .map(|x| x[var_index].to_u32() as u8)
+                .map(|x| x[var_index].to_u32() as u16)
                 .collect();
             c2.coefficients = c.coefficients;
